1. remove admin-web public react logo 2. move auth/type api to info 3. add scala format config
