+++ backend/.scalafmt.conf
@@ -1,9 +1,11 @@
 version = 3.5.8
+maxColumn = 120
 trailingCommas = keep
 docstrings.style = keep
 indent.defnSite = 2
 indent.caseSite = 2
 indent.extendSite = 2
 spaces.inImportCurlyBraces = true
+newlines.usingParamListModifierPrefer = before
 #runner.dialect = scala213
 runner.dialect = scala3
