+++ backend/src/main/scala/com/timzaak/fornet/di/DI.scala
@@ -1,16 +1,16 @@
 package com.timzaak.fornet.di
 
-import com.timzaak.fornet.config.{AppConfig, AppConfigImpl}
+import com.timzaak.fornet.config.{ AppConfig, AppConfigImpl }
 import com.timzaak.fornet.controller.*
 import com.timzaak.fornet.grpc.AuthGRPCController
 import com.timzaak.fornet.mqtt.MqttCallbackController
 import com.timzaak.fornet.mqtt.api.RMqttApiClient
-import com.timzaak.fornet.pubsub.{MqttConnectionManager, NodeChangeNotifyService}
+import com.timzaak.fornet.pubsub.{ MqttConnectionManager, NodeChangeNotifyService }
 import com.timzaak.fornet.service.*
-import com.typesafe.config.{Config, ConfigFactory}
+import com.typesafe.config.{ Config, ConfigFactory }
 import org.hashids.Hashids
-import very.util.keycloak.{JWKPublicKeyLocator, JWKTokenVerifier, KeycloakJWTAuthStrategy}
-import very.util.web.auth.{AuthStrategy, AuthStrategyProvider, SingleUserAuthStrategy}
+import very.util.keycloak.{ JWKPublicKeyLocator, JWKTokenVerifier, KeycloakJWTAuthStrategy }
+import very.util.web.auth.{ AuthStrategy, AuthStrategyProvider, SingleUserAuthStrategy }
 object DI extends DaoDI { di =>
   given config: Config = ConfigFactory.load()
 
@@ -70,10 +70,12 @@ object DI extends DaoDI { di =>
       }
     )
   // web controller
-  object networkController extends NetworkController(
-    networkDao = di.networkDao,
-    appConfig = di.appConfig,
-  )
+  object appInfoController extends AppInfoController(appConfig = di.appConfig)
+  object networkController
+    extends NetworkController(
+      networkDao = di.networkDao,
+      appConfig = di.appConfig,
+    )
   object nodeController
     extends NodeController(
       nodeDao = di.nodeDao,
@@ -107,5 +109,5 @@ object DI extends DaoDI { di =>
       nodeService = di.nodeService,
       mqttConnectionManager = di.connectionManager
     )
-  
+
 }
